Don't retain votes when refreshing page in Firefox
The votes cannot be retained when refreshing, because when loading the page the candidates are placed randomly left or right to remove bias. Reloading the page randomly swaps some candidates but doesn't swap the currently selected candidate.
This change brings the behaviour down to how it works in Chrome: No autocompletion based on votes before the refresh.

For documentation, see:
- https://www.w3schools.com/tags/att_input_autocomplete.asp
- https://stackoverflow.com/questions/7377301/firefox-keeps-form-data-on-reload

Fixes #32.